Adding current iteration of the repl.it files
I downloaded  the replit.zip files for my code and am uploading them as a placeholder for now, as I continue to work on my main.cs. I will be updating that once I finished it on repl.it. I cannot successfully link this repository to repl.it as it is my first time using that platform.